% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{AIC.colsa}
\alias{AIC.colsa}
\title{Calculate the Akaike Information Criterion (AIC) for a 'colsa' object}
\usage{
\method{AIC}{colsa}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"colsa"} for which the AIC is to be
calculated.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A numeric value representing the AIC of the model.
}
\description{
This function computes the AIC for a given 'colsa' object. The AIC is a
measure of the relative quality of a statistical model for a given dataset.
It is calculated as: \eqn{-2 * logLik + 2 * k}, where \code{k} is the number
of parameters in the model.
}
\seealso{
\code{\link{logLik.colsa}}, \code{\link{BIC.colsa}}
}
