% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{basehaz}
\alias{basehaz}
\title{Generic Function for Baseline Hazard}
\usage{
basehaz(object, ...)
}
\arguments{
\item{object}{An object for which the baseline hazard is to be computed.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector representing the baseline hazard.
}
\description{
This is a generic function to compute the baseline hazard for a given object.
}
