% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{Berns_eval}
\alias{Berns_eval}
\title{Evaluate Bernstein Polyomials}
\usage{
Berns_eval(x0, gamma, k, I, J, t, d, X, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Evaluate Bernstein Polyomials
}
