% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{BIC.colsa}
\alias{BIC.colsa}
\title{Calculate Bayesian Information Criterion (BIC) for a 'colsa' Object}
\usage{
\method{BIC}{colsa}(object, ...)
}
\arguments{
\item{object}{An object of class 'colsa'. This is the fitted model for which
the BIC is to be calculated.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A numeric value representing the BIC of the model.
}
\description{
This function computes the Bayesian Information Criterion (BIC) for a fitted
'colsa' model object. The BIC is a model selection criterion that balances
model fit and complexity.
}
\details{
The BIC is calculated as:
\deqn{-2 \times \log(\text{Likelihood}) + k \times \log(n)}
where \eqn{k} is the number of parameters in the model,
and \eqn{n} is the number of samples in the dataset.
}
\seealso{
\code{\link{logLik.colsa}}, \code{\link{AIC.colsa}}
}
