% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{coef.colsa}
\alias{coef.colsa}
\title{Extract Coefficients from a 'colsa' Object}
\usage{
\method{coef}{colsa}(object, ...)
}
\arguments{
\item{object}{An object of class 'colsa'. The fitted model object from which
coefficients are to be extracted.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A numeric vector containing the coefficients extracted from the
'colsa' object.
}
\description{
This function retrieves the coefficients from a fitted 'colsa' object.
}
\details{
The function verifies that the input object is of class 'colsa'.
It then calculates the coefficients by applying a transformation matrix
(\code{prox}) to the \code{theta} parameter of the object. The transformation matrix
accounts for the number of basis functions and features in the model.
}
