% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLSA-package.R
\docType{package}
\name{COLSA-package}
\alias{COLSA}
\alias{COLSA-package}
\title{COLSA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://gongziyang.com/COLSA/}
  \item \url{https://github.com/signorinoy/COLSA}
  \item Report bugs at \url{https://github.com/signorinoy/COLSA/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
