% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{ddloglik_eval.hazard}
\alias{ddloglik_eval.hazard}
\title{Get the Hessian matrix of the likelihood function.}
\usage{
ddloglik_eval.hazard(par, t, d, X, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Get the Hessian matrix of the likelihood function.
}
