% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{density_surv_spline_eval.hazard}
\alias{density_surv_spline_eval.hazard}
\title{Caculate the negative log-likelihood function}
\usage{
density_surv_spline_eval.hazard(par, t, d, X, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Caculate the negative log-likelihood function
}
