% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{dloglik_eval.hazard}
\alias{dloglik_eval.hazard}
\title{Get the first order derivative of the log likelihood function}
\usage{
dloglik_eval.hazard(par, t, d, X, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Get the first order derivative of the log likelihood function
}
