% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{find_inits}
\alias{find_inits}
\title{Finding initial value for COLSA}
\usage{
find_inits(dg, data_first, form, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Finding initial value for COLSA
}
