% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_est_surv}
\alias{get_est_surv}
\title{Get the estimated survival probabilities}
\usage{
get_est_surv(par, t, X, boundaryknots)
}
\arguments{
\item{boundaryknots}{}
}
\description{
Get the estimated survival probabilities
}
