% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{intBerns_eval}
\alias{intBerns_eval}
\title{Numerical integration for the polynomial}
\usage{
intBerns_eval(lb, ub, gamma, k, I, J, t, d, X, breaks)
}
\arguments{
\item{breaks}{}
}
\description{
Numerical integration for the polynomial
}
