% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{logLik.colsa}
\alias{logLik.colsa}
\title{Compute Log-Likelihood for 'colsa' Objects}
\usage{
\method{logLik}{colsa}(object, ...)
}
\arguments{
\item{object}{An object of class 'colsa'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The log-likelihood value stored in the 'colsa' object.
}
\description{
This function retrieves the log-likelihood value from an object of class
'colsa'.
}
