% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prox_forward}
\alias{prox_forward}
\title{Proximal Operator (Forward)}
\usage{
prox_forward(p, q)
}
\arguments{
\item{p}{Integer. The number of basis functions in the original dimension.}

\item{q}{Integer. The number of basis functions in the target dimension.}
}
\value{
A matrix representing the transformation from the original dimension
to the target dimension.
}
\description{
Constructs a transformation matrix to map the coefficients of Bernstein
polynomial basis functions from one dimension to another, ensuring
consistency with the properties of Bernstein polynomial bases. Note that
the original dimension must be less than or equal to the target dimension.
}
\details{
This function iteratively constructs the transformation matrix by
expanding the basis functions from \code{p} to \code{q}, while
preserving the mathematical properties of Bernstein polynomials.
}
\seealso{
\code{\link{prox_reverse}}
}
\keyword{internal}
