% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{summary.colsa}
\alias{summary.colsa}
\title{Summarize a 'colsa' Object}
\usage{
\method{summary}{colsa}(object, conf.int = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"colsa"}.}

\item{conf.int}{A numeric value between 0 and 1 specifying the confidence
level for the confidence intervals. Default is 0.95.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list of class \code{"summary.colsa"} containing the following
components:
\item{call}{The function call that created the \code{"colsa"} object.}
\item{coefficients}{A matrix with columns for the coefficients, exponentiated
coefficients, standard errors, z-scores, and p-values.}
\item{conf.int}{A matrix with columns for the exponentiated coefficients,
their reciprocals, and the lower and upper confidence interval bounds.}
\item{n_basis}{The number of basis functions used in the model.}
\item{boundary}{The boundary information from the \code{"colsa"} object.}
}
\description{
This function provides a summary for objects of class \code{"colsa"}.
}
\details{
The function calculates the z-scores and p-values for the
coefficients, as well as confidence intervals for the exponentiated
coefficients.
}
