% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{Surv}
\alias{Surv}
\title{Alias for Survival Function}
\usage{
Surv(
  time,
  time2,
  event,
  type = c("right", "left", "interval", "counting", "interval2"),
  origin = 0
)
}
\arguments{
\item{time}{The follow-up time for the subject.}

\item{time2}{The ending time for the interval (if applicable).}

\item{event}{The status indicator, normally 0=alive, 1=dead.
Other values are allowed.}

\item{type}{Character string specifying the type of censoring.
Possible values are "right", "left", "interval", or "counting".}

\item{origin}{The origin for counting time, used when \code{type = "counting"}.}
}
\value{
A \code{Surv} object representing the survival data.
}
\description{
This function creates an alias for the \code{Surv} function from the
\code{survival} package, allowing it to be used without explicitly referencing
the package namespace.
}
\seealso{
\code{\link[survival]{Surv}}
}
