% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{update.colsa}
\alias{update.colsa}
\title{Update a COLSA Model with New Data}
\usage{
\method{update}{colsa}(object, newdata, n_basis = "auto", alpha = 1, nu = 0.2, ...)
}
\arguments{
\item{object}{An object of class \code{"colsa"} representing the fitted
COLSA model.}

\item{newdata}{A \code{data.frame} containing the new data to update the
model. It must include the variables specified in the original model
formula.}

\item{n_basis}{Either a numeric value specifying the number of basis
functions or \code{"auto"} to automatically determine the number of basis
functions based on the sample size and scaling parameters.
Defaults to \code{"auto"}, which means the number of basis functions will
be determined by n_basis = round(alpha * n_samples^nu).}

\item{alpha}{A positive numeric value controlling the growth rate of the
number of basis functions when \code{n_basis = "auto"}. Defaults to 1.0.}

\item{nu}{A positive numeric value controlling the scaling exponent for
determining the number of basis functions when \code{n_basis = "auto"}.
Defaults to 0.2.}

\item{...}{Additional arguments passed to the optimization function.}
}
\value{
An updated object of class \code{"colsa"} with the following
components:
\itemize{
\item \code{logLik}: The log-likelihood of the updated model.
\item \code{theta}: The updated parameter estimates.
\item \code{hessian}: The updated Hessian matrix.
\item \code{n_samples}: The updated total number of samples.
\item \code{n_basis}: A vector tracking the number of basis functions used
at each update step.
\item \code{call}: The matched call of the update function.
}
}
\description{
This function updates a COLSA model  with new data, allowing for dynamic
adjustment of the model parameters and basis functions.
}
\details{
The function updates the COLSA model by incorporating new data and
adjusting the number of basis functions dynamically. It ensures that the
model parameters and Hessian matrix are updated accordingly. The
optimization process is performed using the \code{nlm} method to
maximize the log-likelihood of the updated model.
}
