% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLSupdate.R
\name{updateCOLSA.demo}
\alias{updateCOLSA.demo}
\title{COLSA update function for demo}
\usage{
updateCOLSA.demo(B, data, init = NA, p, npar, boundaryknots)
}
\arguments{
\item{B}{number of sites/subdatas}

\item{data}{a dataframe that contains all datasets}

\item{init}{initial beta estimates}

\item{p}{the total number of covaraites}

\item{npar}{the number of covariates in the model}

\item{boundaryknots}{the boundary knots for the Bernstein polynomial}
}
\description{
COLSA update function for demo
}
