% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLSupdate.R
\name{updateCOLSA.outloop}
\alias{updateCOLSA.outloop}
\title{COLSA update function for distributed datasets}
\usage{
updateCOLSA.outloop(subdata, statistics, p, npar, boundaryknots)
}
\arguments{
\item{subdata}{a dataframe that contains the data for a single site}

\item{statistics}{a list that contains the current estimates of beta and Hessian}

\item{p}{the total number of parameters}

\item{npar}{the number of covariates in the model}

\item{boundaryknots}{the boundary knots for the Bernstein polynomial}
}
\description{
COLSA update function for distributed datasets
}
