% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsa.R
\name{vcov.colsa}
\alias{vcov.colsa}
\title{Compute Variance-Covariance Matrix for 'colsa' Objects}
\usage{
\method{vcov}{colsa}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"colsa"}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix representing the variance-covariance matrix of the model
parameters.
}
\description{
This function computes the variance-covariance matrix for objects of class
\code{"colsa"}.
}
